+++ src/rule-definitions.ts
@@ -6,6 +6,7 @@ import forbidRequiredIgnoredField from '#src/rules/forbid-required-ignored-field
 import modelNameGrammaticalNumber from '#src/rules/model-name-grammatical-number.js';
 import modelNameMappingSnakeCase from '#src/rules/model-name-mapping-snake-case.js';
 import modelNamePrefix from '#src/rules/model-name-prefix.js';
+import requireDefaultEmptyArrays from '#src/rules/require-default-empty-arrays.js';
 import requireFieldIndex from '#src/rules/require-field-index.js';
 import requireFieldType from '#src/rules/require-field-type.js';
 import requireField from '#src/rules/require-field.js';
@@ -18,6 +19,7 @@ export default [
   modelNameGrammaticalNumber,
   modelNameMappingSnakeCase,
   modelNamePrefix,
+  requireDefaultEmptyArrays,
   requireField,
   requireFieldIndex,
   requireFieldType,